Refactor PersistentResourceXMLDescription builder into SubsystemSchema implementation.
