+++ clustering/web/extension/src/main/java/org/wildfly/extension/clustering/web/DistributableWebSubsystemSchema.java
@@ -22,9 +22,15 @@
 
 package org.wildfly.extension.clustering.web;
 
+import static org.jboss.as.controller.PersistentResourceXMLDescription.builder;
+
+import java.util.stream.Stream;
+
+import org.jboss.as.clustering.controller.Attribute;
 import org.jboss.as.controller.PersistentResourceXMLDescription;
 import org.jboss.as.controller.PersistentSubsystemSchema;
 import org.jboss.as.controller.SubsystemSchema;
+import org.jboss.as.controller.PersistentResourceXMLDescription.PersistentResourceXMLBuilder;
 import org.jboss.as.controller.xml.VersionedNamespace;
 import org.jboss.staxmapper.IntVersion;
 
@@ -35,8 +41,8 @@ import org.jboss.staxmapper.IntVersion;
 public enum DistributableWebSubsystemSchema implements PersistentSubsystemSchema<DistributableWebSubsystemSchema> {
 
     VERSION_1_0(1, 0), // WildFly 17
-    VERSION_2_0(2, 0), // WildFly 18-26.1
-    VERSION_3_0(3, 0), // WildFly 27
+    VERSION_2_0(2, 0), // WildFly 18-26.1, EAP 7.4
+    VERSION_3_0(3, 0), // WildFly 27-29
     ;
     static final DistributableWebSubsystemSchema CURRENT = VERSION_3_0;
 
@@ -53,6 +59,33 @@ public enum DistributableWebSubsystemSchema implements PersistentSubsystemSchema
 
     @Override
     public PersistentResourceXMLDescription getXMLDescription() {
-        return DistributableWebXMLDescriptionFactory.INSTANCE.apply(this);
+        return builder(DistributableWebResourceDefinition.PATH, this.namespace).addAttributes(Attribute.stream(DistributableWebResourceDefinition.Attribute.class))
+                .addChild(this.getInfinispanSessionManagementResourceXMLBuilder())
+                .addChild(this.getHotRodSessionManagementResourceXMLBuilder())
+                .addChild(builder(InfinispanSSOManagementResourceDefinition.WILDCARD_PATH).addAttributes(Attribute.stream(InfinispanSSOManagementResourceDefinition.Attribute.class)))
+                .addChild(builder(HotRodSSOManagementResourceDefinition.WILDCARD_PATH).addAttributes(Attribute.stream(HotRodSSOManagementResourceDefinition.Attribute.class)))
+                .addChild(builder(LocalRoutingProviderResourceDefinition.PATH).setXmlElementName("local-routing"))
+                .addChild(builder(InfinispanRoutingProviderResourceDefinition.PATH).addAttributes(Attribute.stream(InfinispanRoutingProviderResourceDefinition.Attribute.class)).setXmlElementName("infinispan-routing"))
+                .build();
+    }
+
+    private PersistentResourceXMLBuilder getInfinispanSessionManagementResourceXMLBuilder() {
+        PersistentResourceXMLBuilder builder = builder(InfinispanSessionManagementResourceDefinition.WILDCARD_PATH).addAttributes(Stream.concat(Attribute.stream(InfinispanSessionManagementResourceDefinition.Attribute.class), Attribute.stream(SessionManagementResourceDefinition.Attribute.class)));
+        addAffinityChildren(builder).addChild(builder(PrimaryOwnerAffinityResourceDefinition.PATH).setXmlElementName("primary-owner-affinity"));
+        if (this.namespace.since(DistributableWebSubsystemSchema.VERSION_2_0)) {
+            builder.addChild(builder(RankedAffinityResourceDefinition.PATH).addAttributes(Attribute.stream(RankedAffinityResourceDefinition.Attribute.class)).setXmlElementName("ranked-affinity"));
+        }
+        return builder;
+    }
+
+    private PersistentResourceXMLBuilder getHotRodSessionManagementResourceXMLBuilder() {
+        return addAffinityChildren(builder(HotRodSessionManagementResourceDefinition.WILDCARD_PATH).addAttributes(Stream.concat(Attribute.stream(HotRodSessionManagementResourceDefinition.Attribute.class), Attribute.stream(SessionManagementResourceDefinition.Attribute.class))));
+    }
+
+    private static PersistentResourceXMLBuilder addAffinityChildren(PersistentResourceXMLBuilder builder) {
+        return builder
+                .addChild(builder(NoAffinityResourceDefinition.PATH).setXmlElementName("no-affinity"))
+                .addChild(builder(LocalAffinityResourceDefinition.PATH).setXmlElementName("local-affinity"))
+                ;
     }
 }
